requirements.txt
flask==2.3.3
flask_sqlalchemy==3.1.1
flask_jwt_extended==4.5.3
flask-wtf==1.2.1
flask-limiter==3.5.0
pymysql==1.1.0
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.1
beautifulsoup4==4.12.2
sentence-transformers==2.2.2
scikit-learn==1.3.0
gunicorn==21.2.0
flask-cors==4.0.0
python-json-logger==2.0.7